refactor(ingester): Rename `data_types` module to `dml_payload`
